Typed new files
